Beef up remaining combinator docs, genjax.{mix,switch} signature change (#1174)
This PR:

- **breaking!** Turns `genjax.mix` and `genjax.switch` from decorators
into into functions from `*gen_fn -> gen_fn`. The decorator form was
clunky and not something I want to encourage. Folks can use the methods
if they want `f.mix(g)` , or `mix(f, g)`. But `mix(g)(f)` is no longer a
thing.
- adds documentation and examples for all combinators, decorators and
combinator instance methods

None of the examples have assertions, but these are additional blocks of
code that can prevent breakages, as a bit of a backstop.