Remove AddressBijection + map_addresses, relax decorator type bounds (GEN-322) (#1185)
This PR:

- removes `AddressBijectionCombinator` and the associated
`map_addresses` method/decorator
- removes `AddrMapChm`, `choice_map_address_function`,
`ChoiceMap.with_addr_map`
- implements `RepeatCombinator` without the need for the `"_internal"`
address (and removing our only usage of `map_addresses`
- changes all combinator method / decorator return types to
`GenerativeFunction`. I did this because this gives us more flexibility
to change the implementation later without breaking type compatibility,
and no one should be explicitly relying on the implementation details
anyway

I removed `AddressBijectionCombinator` because it had a few bugs
(mapping `x` to `new_x` left `x` available in the choice map, for
example), and because ironing out the behavior of a combinator that
could flatten addresses was quite hairy.

For example, if we map with

```python
{...: "_ignore",
 "a": "b",
 (): "random"}

```
it seems like this is valid and will
- skip any address one level down except `_ignore`, which will get
flattened
- if I call `get_value()` it’ll get converted to
`.get_submap("random").get_value()`
- `"a"`’s mapping will never be seen by anyone

@femtomc notes that we shouldn't have `...` and `()` in the mapping at
the same time, but we didn't do any validation and this was trouble
waiting to happen imo.