Add `curry` to `StaticGenerativeFunction` (#1247)
This PR adds an instance method `partial_apply` that partially applies
the supplied arguments to the generative function. I'd have called this
`partial`, but Python doesn't allow an instance method to have the same
name as an existing static method.

We need something like this to be able to partially apply generative
functions in a JAX-compatible manner. Calling `Pytree.partial` alone
will not do the right thing.

---------

Co-authored-by: Colin Smith <colin.smith@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>