GEN-482 (fix pyright errors in GenJAX) (#1273)
- pyright was giving different results in (VS Code, pre-commit) vs.
Github action. I adjusted the GitHub action to follow Pylance, as
recommended by the action author, which solved that problem. Previously,
upright version 1.1.358 was in use; following Pylance moves that version
to 1.1.373.
- Sample is replaced with ChoiceMap at all points where type errors
follow the use of the former. I understand that ChoiceMap is a narrower
concept than Sample, but we do not have enough type infrastructure in
GenJAX to support the distinction presently
- the `inference` library is temporarily excluded, as it is not ready
for prime-time
- the handling of static booleans is now handled by `Flag`. The point of
`Flag` is to concentrate the logic for making concreteness optimizations
in one place. I believe that the systematic use of Flag has corrected a
few spots where concreteness optimization had been overlooked, provoking
the elision of a few `unmask` calls in the unit test suite. These
require special attention to make sure that my conjecture is true.
- Constructing a Mask requires a Flag. It may be better to have the Mask
constructor promote bare booleans to Flag types, or, to add a
`post_init` assertion that the type is correct. It may be, however, that
the user shouldn't be creating their own Masks, and so we can rely on
the type checking for library development from now on to enforce correct
Mask creation.
- pyright is (perhaps overly) scrupulous about the names of parameters
in overridden methods, and doesn't allow overriding `*args` with a
sequence of better names. Can't fight city hall, so I have changed this.
- To clear some errors I have resorted to a pattern of `assert
isinstance(x, Foo), f"A Foo is required here, not {x}, [because ...]"`,
which I think we should adopt since it's harder to figure out what to do
if beartype gets involved.
-

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Sam Ritchie <sam@mentat.org>