Remove reverse, unroll flags from scan and friends (GEN-404) (#1328)
This PR removes the reverse and unroll options that currently flow
through to `jax.lax.scan`. @femtomc and I did some investigation and
found that the behavior here is weird enough that we don't want to take
on interpreting it, given that no one yet wants this and no one is using
it!

`reverse=True` reverses the input sequence, but ALSO reverses the
accumulated output. The confusion here is, what do the indices mean in
the ChoiceMap?

If we do take on support of `reverse`, we should copy the `keep_order`
argument here and re-reverse the accumulated outputs:
https://flax.readthedocs.io/en/latest/_modules/flax/linen/recurrent.html#RNN
In

Properly supporting reverse means explaining to the users that we are
walking from right to left on the input array, and telling a good story
about what indices in the choicemap refer to what entries in the output.