Remove ellipsis support from queries, add `SwitchChm` (GEN-698, GEN-696, GEN-542) (#1421)
This PR:

- renames `Mask.maybe` to `Mask.build` and `Mask.maybe_none` to
`Mask.maybe_mask`, better docstrings throughout
- adds an `__or__` and `__xor__` implementation to `Mask`; this is now
used by the `Or` and `Xor` choicemap classes instead of their
`get_value` implementations
- adds a `Switch` subclass of `ChoiceMap`, which is now returned by
`ChoiceMap.switch`
- adds a `flatten` method to `Mask` that drops the value to `None` or
`.value` in the case of concrete flags
- removes `genjax._src.core.typing.Value`; this is just `Any`, let's
just say it.
- replaces all `...` lookups with `:`, since that's what we are actually
going for
- adds a constructor optimization to `Choice`; around a concrete `False`
mask the choice simplifies to `None`, and around a concrete `True` mask
the mask is stripped off.
- removes support for ellipses in `__getitem__`, `__call__` and
`__contains__` for `Selection` and `ChoiceMap`.

(The reason for this is that `ChoiceMap.__getitem__` with a wildcard
forces us to `xor` together the values for all possible keys. Users can
now use `:` in place of `...` for `Indexed` layers, and we don't need
the confusing wildcard lookup for a static level.)

## Next steps:

I am moving toward deleting `Or` and `Xor`, but `Filtered` is currently
in the way. The next step is to see if we can push filters down to the
bottom, and delete the `Filtered` class. See #1422