Enforce "Mask has scalar flag", allow vectorized `Mask` (GEN-768, GEN-765) (#1423)
This PR:

- adds a `Mask.__init__` that checks that 
  - a `Mask` is not instantiated wrapping another `Mask`
- the `Mask` has either a scalar flag, or if the flag has a shape, its
shape is a prefix of all leaves of the value. This check ensures that we
can treat the `Mask` as a vectorized structure and `vmap` down to the
point where we have a scalar `Flag` again.
- modifies `maybe_mask` to preserve masked `None`s. This method should
only be looking at the flag!
- adds an `__invert__` implementation to `Mask`, allowing `~mask`
- when calling `|` or `^` between two masks, validates that the shapes
of the values and flags match. This ensures that we aren't combining
`Mask`s at different vectorization levels
- when called with a `Mask` value, `Mask.build` has to either take a
scalar flag OR a flag that matches the already-vectorized shape of the
`Mask` value
-  checks that `or` and `xor` work between vectorized values
- modifies all of the `FlagOp` operations so that an `Array` on EITHER
side coerces the result into an `Array`. We need this to maintain our
vectorization levels, so that a scalar `bool` flag can only live in a
non-vectorized `Mask`
- adds type annotations helping the compiler understand this latest
check