Fix assess in vmap (GEN-903) (#1464)
This PR modifies `assess` in vmap to vmap over indices and query the
choicemap for each, rather than vmapping over the `ChoiceMap` itself.

The error @georgematheos was hitting was that his choicemap had masks
inside where a vmapped value was masked with a scalar bool. It's not
possible to vmap over this because the scalar bool doesn't have the same
leading dimension.