GEN-957 internal `StaticTrace` improvements (#1500)
- AddressVisitor is eliminated. Since 3.7, Python guarantees iteration
order == insertion order in dictionaries. This means having a list of
visited addresses together with a vector of traces at those addresses is
not necessary: a dictionary can serve this purpose.
- Adds a VoidTrace object to stash in such dictionaries to defend
against duplicate traces
- These new subtrace dictionaries store subtrace names as bare strings
when that is the type of the trace address (i.e., almost always). This
eliminates the annoying `get_subtrace((v,))` pattern. This usage is
still supported, but now provokes a deprecation warning.
- This means there is no further confusion about what a tuple in a trace
list means: the address is the address, so if you trace to a tuple
address, the tuple enters the dictionary. This allows the collapse of
several instances of code that shifted between string and 1-tuple
representation addresses.
- In consequence of all this, the type definition of StaticAddress is
now `StaticAddress = StaticAddressComponent |
tuple[StaticAddressComponent, ...]`, which is a big change, but I think
reflective of how addresses are actually used in GenJAX.