Expunge the concept of Constraint; just rely on choice map everywhere. (#1532)
This removes `Constraint` from the core set of concepts, simplifying
several of the types in interface signatures (this is similar to the PR
which expunged `Sample`).

In the future, when special assumptions about constraints need to be
communicated, I expect these to live directly in `EditRequest` or (to be
introduced), `GenerateRequest` (corresponding to the "advanced"
interfaces `edit` ("change an existing trace") and `generate` ("sample a
new properly weighted trace")).

---------

Co-authored-by: Sam Ritchie <sam@mentat.org>