Fix mix side-effecting issue (GEN-986) (#1537)
This PR moves the `@gen` decorator down to a `gen(...)` call. Before
doing this, beartype 0.18.0 would somehow capture the `mixture_model`
function internally in such a way that calls to `simulate` (with nested
models only, not single layer models) of a totally different model would
try and call `mixture_model`.

This change gets around that issue.