Add support for sample_shape to primitive distributions (#1576)
This PR passes any `sample_shape` argument provided to a tfp
distribution through to the `sample` method, instead of passing it to
the constructor.

@femtomc one issue that I hit here was that to use this, I have to wrap
the argument in `Const`, because it seems like our code tries to trace
non-jit-compiled fns, or something like that. I'll add the error I saw
as a reply.

Co-authored-by: Mathieu Huot <MathieuHuot@users.noreply.github.com>